[Mosaic GPU] Remove unused `vector.load` before layout inference.
There are cases where layout inference fails with such loads. This CL adds a pass to remove these. It's not that nice, but we are thinking of rewriting this logic anyway, so perhaps it's an OK hack. I tried running dce and/or canonicalize instead of this custom pass, but those didn't do anything. Perhaps they see that the load has side-effects. Open to other suggestions.

PiperOrigin-RevId: 748323295